Adjust unit tests for new op_type and Python 3.6
This PR changes the indexing op_type to the default `index`, so test fixtures
needed to be adjusted to reflect that.

This also adjusts the tests to run in Python 3.6 instead of 3.7, since that's
the version that will execute the actual pipeline, and I also changed the
test chunk size to 2000 to match usage in production.

NOTE:
Tests will still fail for linting, but that has been addressed in main.