Remove code that built the decommissioned landing-page hero map
The hero map was removed in February 2022, but the code that generated
the map data was still churning away every day in Jenkins.

The Jenkins job is still needed to generate other metadata, so the way to
turn off the hero-map work is to remove it from this repo's Makefile and
from the complaints/ccdb folder.

This also cleans up some README language and removes an unneeded declaration
in common/es_proxy.py