Bump the indexing timeout value
Indexing on EXT Jenkins has become troublesome, possibly because of the
increased load on the server from more aggressive security scanning and
heavy morning cron jobs.

Two types of errors have interrupted indexing:
- Disconnects with S3
- Timeouts during OpenSearch indexing

These appear to be specific to EXT Jenkins, because our DEV Jenkins instance
(zusa) runs the same pipeline code against the same data, at the same time
of the morning, and rarely fails to finish.

The S3 interruptions are not too disruptive, because the downloads that succeed
won't have to be run again on a retry, and little time is lost.

The timeout error, however, is painful when indexing bombs late in the indexing run.
In the last week, a timeout stopped indexing after 4 million complaints had been
indexed, which causes the job to start over from 0.

This PR doubles the OpenSearch timeout value, which should not affect most runs, but could save the occasional late timeout.

I ran this morning's CCDB indexing using this branch, and it succeeded on the first try.
That doesn't prove that the new value saved the run, but I think we should see if the new timeout reduces the churn.