Add a '--narratives' flag for local indexing
When passed to the csv2json function, the flag will index only complaints
with narratives. This helps when testing locally, because only 29% of CCDB
complaints have narratives, and if you index a small slice
of our 12M complaints, you might end up with no narratives.

This can be jarring when navigating to the CCDB page from the
"Read complaints" link on the complaints landing page
(/data-research/consumer-complaints/). That link
sets the 'has_narratives=true' filter, and no complaints will show.